refactor: extract ToastIcon/CloseButton; single-pass position derivation
Address React Doctor findings: toast.tsx declared three components
(no-multi-comp) — ToastIcon and CloseButton move verbatim to their own
files; the positionsData memo combined .filter().map() into one
for...of pass (js-combine-iterations) with identical container
semantics (default position always renders, others only with members).
