+++ src/toast-icon.tsx
@@ -1,0 +1,25 @@
+import * as React from 'react';
+import { CircleCheck, CircleX, Info, TriangleAlert } from './icons';
+import type { ToastProps } from './types';
+import { useIconColor } from './use-default-styles';
+
+export const ToastIcon: React.FC<
+  Pick<ToastProps, 'variant'> & {
+    invert: boolean;
+    richColors: boolean;
+  }
+> = ({ variant, invert, richColors }) => {
+  const color = useIconColor({ variant, invert, richColors });
+
+  switch (variant) {
+    case 'success':
+      return <CircleCheck size={20} color={color} />;
+    case 'error':
+      return <CircleX size={20} color={color} />;
+    case 'warning':
+      return <TriangleAlert size={20} color={color} />;
+    default:
+    case 'info':
+      return <Info size={20} color={color} />;
+  }
+};
